+++ src/main/java/com/qindesign/json/schema/JSON.java
@@ -39,6 +39,9 @@ import java.io.InputStreamReader;
 import java.io.Reader;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * Provides JSON tools, including:
@@ -173,10 +176,10 @@ public final class JSON {
      *
      * @param e the element being visited
      * @param parent the element's parent, may be {@code null}
-     * @param path the full path to the element, a JSON Pointer
+     * @param path the full path to the element, a list of path elements
      * @param state holds more information about the element
      */
-    void visit(JsonElement e, JsonElement parent, String path, TraverseState state);
+    void visit(JsonElement e, JsonElement parent, List<String> path, TraverseState state);
   }
 
   /**
@@ -214,19 +217,21 @@ public final class JSON {
    * @param visitor the visitor
    */
   public static void traverse(JsonElement e, JsonElementVisitor visitor) {
-    traverse(e, null, "", new TraverseState(), visitor);
+    traverse(e, null, new ArrayList<>(), new TraverseState(), visitor);
   }
 
   /**
    * Recursive method that performs the traversal.
+   * <p>
+   * The path is expected to be a modifiable {@link List}.
    *
    * @param e the current element
    * @param parent the element's parent
-   * @param path the element's full path, a JSON Pointer
+   * @param path the element's full path, a list of path elements
    * @param state the tree state
    * @param visitor the visitor
    */
-  private static void traverse(JsonElement e, JsonElement parent, String path,
+  private static void traverse(JsonElement e, JsonElement parent, List<String> path,
                                TraverseState state,
                                JsonElementVisitor visitor) {
     // Possibly alter the state
@@ -236,26 +241,27 @@ public final class JSON {
       state.spec = spec;
     }
 
-    visitor.visit(e, parent, path, state);
+    visitor.visit(e, parent, Collections.unmodifiableList(new ArrayList<>(path)), state);
 
     if (e.isJsonPrimitive() || e.isJsonNull()) {
       return;
     }
 
+    path.add("");
     if (e.isJsonArray()) {
       int index = 0;
       for (var item : e.getAsJsonArray()) {
-        traverse(item, e, path + "/" + index, state, visitor);
+        path.set(path.size() - 1, Integer.toString(index));
+        traverse(item, e, path, state, visitor);
         index++;
       }
-      return;
+    } else {
+      // Process everything else
+      for (var entry : e.getAsJsonObject().entrySet()) {
+        path.set(path.size() - 1, entry.getKey());
+        traverse(entry.getValue(), e, path, state, visitor);
+      }
     }
-
-    // Process everything else
-    for (var entry : e.getAsJsonObject().entrySet()) {
-      traverse(entry.getValue(), e,
-               path + "/" + Strings.jsonPointerToken(entry.getKey()), state,
-               visitor);
-    }
+    path.remove(path.size() - 1);
   }
 }